matrix format and metadata split
