--- input/aa_changes.txt
@@ -1,97 +1,0 @@
-S:L5F
-S:L18F
-S:T19R
-S:T19I
-S:T20N
-S:del:24:4
-S:P26S
-S:T29A
-S:W64R
-S:A67V
-S:del:68:3
-S:T95I
-S:G142D
-S:del:142:4
-S:del:143:2
-S:Y145H
-S:H146Q
-S:del:156:3
-S:R190S
-S:del:211:2
-S:V213G
-S:R214REPE
-S:D215G
-S:A222V
-S:del:240:4
-S:T250I
-S:T259A
-S:T299I
-S:G339D
-S:G339H
-S:R346K
-S:R346T
-S:L368I
-S:S371L
-S:S371F
-S:S373P
-S:S375F
-S:T376A
-S:D405N
-S:R408S
-S:K417N
-S:K417T
-S:N440K
-S:K444T
-S:V445P
-S:G446S
-S:S477N
-S:L452R
-S:L452Q
-S:L452M
-S:T478K
-S:E484K
-S:E484Q
-S:E484A
-S:F486V
-S:F490V
-S:Q493R
-S:G496S
-S:Q498R
-S:N501Y
-S:Y505H
-S:T547K
-S:A570D
-S:Q613H
-S:D614G
-S:H655Y
-S:Q677H
-S:N679K
-S:P681H
-S:P681R
-S:A701V
-S:S704L
-S:T716I
-S:N764K
-S:D796Y
-S:N856K
-S:D950N
-S:Q954H
-S:N969K
-S:L981F
-S:S982A
-S:T1027I
-S:D1118H
-S:V1176F
-S:S371L S:S373P
-S:del:68:3 S:N501Y S:K417N
-S:del:68:3 S:E484K
-S:del:68:3 S:E484A
-S:N501Y S:K417N
-S:N501Y S:K417T
-S:N501Y S:H655Y
-S:del:68:3 S:N501Y
-S:T29A S:T250I S:T299I S:Q613H
-S:Y145H S:A222V
-ORF3a:H78Y
-N:S33F
-N:P151S
