--- input/signature.txt
@@ -1,7 +1,0 @@
-BA.2.75.2
-BA.4.6
-BA.5
-BF.7
-BQ.1
-XBB.1
-XBB.1.1
