--- input/voi_lineages_2023-04-17_2023-06-12.txt
@@ -1,27 +1,0 @@
-VOI,XBB.2.3,Omicron
-VOI,XBB.1.9.2,Omicron
-VOI,XBB.1.9.1,Omicron
-VOI,XBB.1.9,Recombinant
-VOI,XBB.1.5,Recombinant
-VOI,XBB.1.5.60,Recombinant
-VOI,XBB.1.5.37,Recombinant
-VOI,XBB.1.5.65,Recombinant
-VOI,XBB.1.5.13,Recombinant
-VOI,XBB.1.5.28,Recombinant
-VOI,XBB.1.5.7,Recombinant
-VOI,XBB.1.5.62,Recombinant
-VOI,XBB.1.5.20,Recombinant
-VOI,XBB.1.5.38,Recombinant
-VOI,XBB.1.5.21,Recombinant
-VOI,XBB.1.5.12,Recombinant
-VOI,XBB.1.5.9,Recombinant
-VOI,XBB.1.5.14,Recombinant
-VOI,XBB.1.16,Recombinant
-VOI,XBB,Recombinant
-VOI,EG.1,Omicron
-VOI,CH.1.1,Omicron
-VOI,BQ.1,Omicron
-VOI,BQ.1.1,Omicron
-VOI,BA.5,Omicron
-VOI,BA.4,Omicron
-VOI,BA.2.75.2,Omicron
